Changed one method of lxd utils to brctl
